portpy/ai/requirements.txt
dominate>=2.6.0
torch>=1.9.1+cu111
torchfile>=0.1.0
torchvision>=0.9.1
typing-extensions==3.10.0.0
urllib3>=1.26.5
visdom>=0.1.8.9
websocket-client>=1.0.1
numpy>=1.21.6
Pillow>=9.5.0
SimpleITK>=2.2.1
matplotlib>=3.5.3
pandas>=1.3.5
pydicom>=2.4.1
scikit-image>=0.19.3